refactor(destregistry): extract NewHTTPClient as free function
MakeHTTPClient never touched BaseProvider state — it was a free helper
pretending to be a method. Move it to httpclient.go and update the three
HTTP-based providers (desthookdeck, destwebhook, destwebhookstandard) to
call destregistry.NewHTTPClient directly. BaseProvider keeps only the
metadata/validation concerns it actually owns.

No behavior change.
