+++ internal/destregistry/providers/destwebhook/destwebhook.go
@@ -342,7 +342,7 @@ func (d *WebhookDestination) CreatePublisher(ctx context.Context, destination *m
 		proxyURL = &d.proxyURL
 	}
 
-	httpClient, err := d.BaseProvider.MakeHTTPClient(destregistry.HTTPClientConfig{
+	httpClient, err := destregistry.NewHTTPClient(destregistry.HTTPClientConfig{
 		UserAgent: &d.userAgent,
 		ProxyURL:  proxyURL,
 	})
